Move loose functions in test/api.js to test/util.js
