+++ birds/test/util.js
@@ -1,0 +1,20 @@
+var exports = module.exports = {};
+
+exports.clone = (obj) => {
+    if (null == obj || "object" != typeof obj) return obj;
+    var copy = obj.constructor();
+    for (var attr in obj) {
+        if (obj.hasOwnProperty(attr)) copy[attr] = obj[attr];
+    }
+    return copy;
+}
+
+exports.equal = (obj1, obj2) => {
+    return JSON.stringify(obj1) == JSON.stringify(obj2);
+}
+
+exports.randomTeamNumber = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min)) + min;
+}
